# design_docs/openapi.v0.yaml
openapi: 3.0.3
info:
  title: Curious Now API (v0)
  version: "0.1"
servers:
  - url: /v1
paths:
...
  /feedback:
    post:
      summary: Stage 8 user feedback (auth optional)
      security:
        - SessionAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FeedbackIn" }
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackResponse" }
...
  /admin/feedback:
    get:
      summary: Stage 8 feedback triage queue
      security:
        - AdminTokenAuth: []
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [new, triaged, resolved, ignored]
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Feedback reports
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminFeedbackListResponse" }
  /admin/feedback/{id}:
    patch:
      summary: Stage 8 update feedback status
      security:
        - AdminTokenAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminFeedbackPatchRequest" }
      responses:
        "200":
          description: Updated feedback report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackReport" }
  /admin/clusters/{id}/merge:
    post:
      summary: Stage 8 merge a cluster into another (creates redirect)
      security:
        - AdminTokenAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminClusterMergeRequest" }
      responses:
        "200":
          description: Merge completed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminClusterMergeResponse" }
  /admin/clusters/{id}/split:
    post:
      summary: Stage 8 split a cluster by extracting items into a new cluster
      security:
        - AdminTokenAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminClusterSplitRequest" }
      responses:
        "200":
          description: Split completed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminClusterSplitResponse" }
  /admin/clusters/{id}/quarantine:
    post:
      summary: Stage 8 quarantine a cluster (stop auto attachments)
      security:
        - AdminTokenAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminClusterQuarantineRequest" }
      responses:
        "200":
          description: Quarantined
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOkResponse" }
  /admin/clusters/{id}/unquarantine:
    post:
      summary: Stage 8 unquarantine a cluster
      security:
        - AdminTokenAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminClusterQuarantineRequest" }
      responses:
        "200":
          description: Unquarantined
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOkResponse" }
  /admin/clusters/{id}:
    patch:
      summary: Stage 8 manual cluster overrides (citations required)
      security:
        - AdminTokenAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminClusterPatchRequest" }
      responses:
        "200":
          description: Updated cluster
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClusterDetail" }
  /admin/clusters/{id}/topics:
    put:
      summary: Stage 8 set cluster topics (manual + locked)
      security:
        - AdminTokenAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminSetClusterTopicsRequest" }
      responses:
        "200":
          description: Updated cluster topics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOkResponse" }
  /admin/topics:
    post:
      summary: Stage 8 create a topic
      security:
        - AdminTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminTopicCreateRequest" }
      responses:
        "200":
          description: Created topic
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Topic" }
  /admin/topics/{id}:
    patch:
      summary: Stage 8 update a topic (name, description, aliases)
      security:
        - AdminTokenAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminTopicPatchRequest" }
      responses:
        "200":
          description: Updated topic
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Topic" }
  /admin/topics/{id}/merge:
    post:
      summary: Stage 8 merge a topic into another (creates redirect)
      security:
        - AdminTokenAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminTopicMergeRequest" }
      responses:
        "200":
          description: Merge completed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminTopicMergeResponse" }
  /admin/lineage/nodes:
    post:
      summary: Stage 8 create a lineage node (paper/model/dataset/method)
      security:
        - AdminTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminLineageNodeCreateRequest" }
      responses:
        "200":
          description: Created lineage node
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LineageNode" }
  /admin/lineage/edges:
    post:
      summary: Stage 8 create a lineage edge (evidence required)
      security:
        - AdminTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminLineageEdgeCreateRequest" }
      responses:
        "200":
          description: Created lineage edge
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LineageEdge" }
...
components:
  securitySchemes:
    AdminTokenAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
    SessionAuth:
      type: apiKey
      in: cookie
      name: cn_session
  schemas:
    ContentType:
      type: string
      enum: [news, press_release, preprint, peer_reviewed, report]
...
    TopicChip:
      type: object
      required: [topic_id, name, score]
      properties:
        topic_id: { type: string, format: uuid }
        name: { type: string }
        score: { type: number }
...
    EvidenceItem:
      type: object
      required: [item_id, title, url, source, published_at]
      properties:
        item_id: { type: string, format: uuid }
        title: { type: string }
        url: { type: string }
        published_at: { type: string, format: date-time, nullable: true }
        source:
          type: object
          required: [source_id, name]
          properties:
            source_id: { type: string, format: uuid }
            name: { type: string }
        content_type: { $ref: "#/components/schemas/ContentType" }
    ClusterDetail:
      type: object
      required: [cluster_id, canonical_title, created_at, updated_at, distinct_source_count, evidence]
      properties:
        cluster_id: { type: string, format: uuid }
        canonical_title: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        distinct_source_count: { type: integer }
        topics:
          type: array
          items: { $ref: "#/components/schemas/TopicChip" }
        content_type_breakdown:
          type: object
          additionalProperties: { type: integer }
        evidence:
          type: object
          additionalProperties:
            type: array
            items: { $ref: "#/components/schemas/EvidenceItem" }
        takeaway: { type: string, nullable: true }
        summary_intuition: { type: string, nullable: true }
        summary_deep_dive: { type: string, nullable: true }
        assumptions:
          type: array
          items: { type: string }
        limitations:
          type: array
          items: { type: string }
        what_could_change_this:
          type: array
          items: { type: string }
        confidence_band:
          type: string
          nullable: true
          enum: [early, growing, established]
        method_badges:
          type: array
          items: { type: string }
        anti_hype_flags:
          type: array
          items: { type: string }
        takeaway_supporting_item_ids:
          type: array
          items: { type: string, format: uuid }
        summary_intuition_supporting_item_ids:
          type: array
          items: { type: string, format: uuid }
        summary_deep_dive_supporting_item_ids:
          type: array
          items: { type: string, format: uuid }
        glossary_entries:
          type: array
          items: { $ref: "#/components/schemas/GlossaryEntry" }
        is_saved: { type: boolean, nullable: true }
        is_watched: { type: boolean, nullable: true }
    Topic:
      type: object
      required: [topic_id, name]
      properties:
        topic_id: { type: string, format: uuid }
        name: { type: string }
        description_short: { type: string, nullable: true }
        is_followed: { type: boolean, nullable: true }
...
    LineageNode:
      type: object
      required: [node_id, node_type, title]
      properties:
        node_id: { type: string, format: uuid }
        node_type:
          type: string
          enum: [paper, model, dataset, method]
        title: { type: string }
        external_url: { type: string, nullable: true }
        published_at: { type: string, format: date-time, nullable: true }
    LineageEdge:
      type: object
      required: [from, to, relation_type, evidence_item_ids]
      properties:
        from: { type: string, format: uuid }
        to: { type: string, format: uuid }
        relation_type:
          type: string
          enum: [extends, improves, compresses, replaces_in_some_settings, contradicts, orthogonal]
        evidence_item_ids:
          type: array
          items: { type: string, format: uuid }
        notes_short: { type: string, nullable: true }
...
    GlossaryEntry:
      type: object
      required: [glossary_entry_id, term, definition_short]
      properties:
        glossary_entry_id: { type: string, format: uuid }
        term: { type: string }
        definition_short: { type: string }
        definition_long: { type: string, nullable: true }
...
    SimpleOkResponse:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [ok] }
...
    FeedbackIn:
      type: object
      required: [feedback_type]
      properties:
        feedback_type:
          type: string
          enum: [confusing, overstated, incorrect, missing_context, broken_link, other]
        client_id: { type: string, format: uuid, nullable: true }
        cluster_id: { type: string, format: uuid, nullable: true }
        item_id: { type: string, format: uuid, nullable: true }
        topic_id: { type: string, format: uuid, nullable: true }
        message: { type: string, nullable: true }
    FeedbackResponse:
      type: object
      required: [status, feedback_id]
      properties:
        status: { type: string, enum: [accepted] }
        feedback_id: { type: string, format: uuid }
    FeedbackReport:
      type: object
      required: [feedback_id, created_at, feedback_type, status]
      properties:
        feedback_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        feedback_type:
          type: string
          enum: [confusing, overstated, incorrect, missing_context, broken_link, other]
        status:
          type: string
          enum: [new, triaged, resolved, ignored]
        message: { type: string, nullable: true }
        cluster_id: { type: string, format: uuid, nullable: true }
        item_id: { type: string, format: uuid, nullable: true }
        topic_id: { type: string, format: uuid, nullable: true }
        user_id: { type: string, format: uuid, nullable: true }
    AdminFeedbackListResponse:
      type: object
      required: [page, results]
      properties:
        page: { type: integer }
        results:
          type: array
          items: { $ref: "#/components/schemas/FeedbackReport" }
    AdminFeedbackPatchRequest:
      type: object
      properties:
        status:
          type: string
          enum: [triaged, resolved, ignored]
        resolution_notes: { type: string, nullable: true }
    AdminClusterMergeRequest:
      type: object
      required: [to_cluster_id]
      properties:
        to_cluster_id: { type: string, format: uuid }
        notes: { type: string, nullable: true }
        supporting_item_ids:
          type: array
          items: { type: string, format: uuid }
    AdminClusterMergeResponse:
      type: object
      required: [from_cluster_id, to_cluster_id]
      properties:
        from_cluster_id: { type: string, format: uuid }
        to_cluster_id: { type: string, format: uuid }
    AdminClusterSplitRequest:
      type: object
      required: [move_item_ids]
      properties:
        new_cluster_title: { type: string, nullable: true }
        move_item_ids:
          type: array
          items: { type: string, format: uuid }
        notes: { type: string, nullable: true }
    AdminClusterSplitResponse:
      type: object
      required: [source_cluster_id, new_cluster_id]
      properties:
        source_cluster_id: { type: string, format: uuid }
        new_cluster_id: { type: string, format: uuid }
    AdminClusterQuarantineRequest:
      type: object
      properties:
        notes: { type: string, nullable: true }
    AdminClusterPatchRequest:
      type: object
      properties:
        canonical_title: { type: string }
        representative_item_id: { type: string, format: uuid, nullable: true }
        takeaway: { type: string, nullable: true }
        takeaway_supporting_item_ids:
          type: array
          items: { type: string, format: uuid }
        summary_intuition: { type: string, nullable: true }
        summary_intuition_supporting_item_ids:
          type: array
          items: { type: string, format: uuid }
        summary_deep_dive: { type: string, nullable: true }
        summary_deep_dive_supporting_item_ids:
          type: array
          items: { type: string, format: uuid }
        assumptions:
          type: array
          items: { type: string }
        limitations:
          type: array
          items: { type: string }
        what_could_change_this:
          type: array
          items: { type: string }
        confidence_band:
          type: string
          nullable: true
          enum: [early, growing, established]
        method_badges:
          type: array
          items: { type: string }
        anti_hype_flags:
          type: array
          items: { type: string }
    AdminSetClusterTopicsRequest:
      type: object
      required: [topics]
      properties:
        replace:
          type: boolean
          default: true
        topics:
          type: array
          items: { $ref: "#/components/schemas/AdminClusterTopicAssignment" }
        notes: { type: string, nullable: true }
    AdminClusterTopicAssignment:
      type: object
      required: [topic_id]
      properties:
        topic_id: { type: string, format: uuid }
        score: { type: number, nullable: true }
        locked: { type: boolean, nullable: true }
    AdminTopicCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description_short: { type: string, nullable: true }
        aliases:
          type: array
          items: { type: string }
        parent_topic_id: { type: string, format: uuid, nullable: true }
    AdminTopicPatchRequest:
      type: object
      properties:
        name: { type: string }
        description_short: { type: string, nullable: true }
        aliases:
          type: array
          items: { type: string }
        parent_topic_id: { type: string, format: uuid, nullable: true }
    AdminTopicMergeRequest:
      type: object
      required: [to_topic_id]
      properties:
        to_topic_id: { type: string, format: uuid }
        notes: { type: string, nullable: true }
    AdminTopicMergeResponse:
      type: object
      required: [from_topic_id, to_topic_id]
      properties:
        from_topic_id: { type: string, format: uuid }
        to_topic_id: { type: string, format: uuid }
    AdminLineageNodeCreateRequest:
      type: object
      required: [node_type, title]
      properties:
        node_type:
          type: string
          enum: [paper, model, dataset, method]
        title: { type: string }
        external_url: { type: string, nullable: true }
        published_at: { type: string, format: date-time, nullable: true }
        external_ids:
          type: object
          additionalProperties: true
        topic_ids:
          type: array
          items: { type: string, format: uuid }
    AdminLineageEdgeCreateRequest:
      type: object
      required: [from_node_id, to_node_id, relation_type, evidence_item_ids]
      properties:
        from_node_id: { type: string, format: uuid }
        to_node_id: { type: string, format: uuid }
        relation_type:
          type: string
          enum: [extends, improves, compresses, replaces_in_some_settings, contradicts, orthogonal]
        evidence_item_ids:
          type: array
          items: { type: string, format: uuid }
        notes_short: { type: string, nullable: true }